fix(copilot): preserve custom thread titles
